Undated stylistic+export changes
