+++ sktime/regression/deep_learning/lstmfcn/__init__.py
@@ -1,0 +1,13 @@
+"""LSTM-FCN deep learning regressors.
+
+This subpackage provides LSTM-FCN based regressors implemented in
+TensorFlow and PyTorch backends.
+"""
+
+from sktime.regression.deep_learning.lstmfcn._lstmfcn_tf import LSTMFCNRegressorTF
+from sktime.regression.deep_learning.lstmfcn._lstmfcn_torch import LSTMFCNRegressorTorch
+
+__all__ = [
+    "LSTMFCNRegressorTorch",
+    "LSTMFCNRegressorTF",
+]
